fix navbar and headline section icon display
