+++ src/index.html
@@ -27,7 +27,7 @@
         </li>
         <li class="section-link envelope">
           <a href="#">
-            <img src="./icons/Envelope-Icon.svg" alt="mail icon" />
+            <img class="envelope-icon" src="./icons/Envelope-Icon.svg" alt="mail icon" />
           </a>
         </li>
       </ul>
@@ -56,27 +56,27 @@
       <ul class="headline-social-media">
         <li>
           <a href="#">
-            <img src="./icons/Disabledgithub-icon.png" alt="gitHub icon" />
+            <img src="" alt="gitHub icon" />
           </a>
         </li>
         <li>
           <a href="#">
-            <img src="./icons/DisabledLinkedIn-icon.png" alt="LinkedIn Icon" />
+            <img src="" alt="LinkedIn Icon" />
           </a>
         </li>
         <li>
           <a href="#">
-            <img src="./icons/Enabledtwo-fingers-icon.png" alt="Two fingers Icons" />
+            <img src="" alt="Two fingers Icons" />
           </a>
         </li>
         <li>
           <a href="#">
-            <img src="./icons/Disabledtwitter-icon.png" alt="Twitter Icon" />
+            <img src="" alt="Twitter Icon" />
           </a>
         </li>
         <li>
           <a href="#">
-            <img src="./icons/DisabledM-icon.png" alt="Make Icon" />
+            <img src="" alt="Make Icon" />
           </a>
         </li>
       </ul>
